--- app/src/main/java/com/topjohnwu/magisk/di/MiscModule.kt
@@ -1,10 +1,0 @@
-package com.topjohnwu.magisk.di
-
-import org.koin.dsl.module
-
-
-val miscModule = module {
-
-    // define miscs here
-
-}
