+++ app/src/main/java/com/topjohnwu/magisk/di/Modules.kt
@@ -5,6 +5,5 @@ val koinModules = listOf(
     networkingModule,
     databaseModule,
     repositoryModule,
-    viewModelModules,
-    miscModule
+    viewModelModules
 )
