--- app/src/main/java/com/topjohnwu/magisk/di/NamedInjection.kt
@@ -1,6 +1,0 @@
-package com.topjohnwu.magisk.di
-
-import org.koin.core.qualifier.named
-
-val SUTimeout = named("su_timeout")
-val Protected = named("protected")
