+++ app/src/main/java/com/topjohnwu/magisk/model/events/ViewEvents.kt
@@ -2,7 +2,6 @@ package com.topjohnwu.magisk.model.events
 
 import android.app.Activity
 import com.skoumal.teanity.viewevents.ViewEvent
-import com.topjohnwu.magisk.model.entity.Policy
 import com.topjohnwu.magisk.model.entity.module.Repo
 import io.reactivex.subjects.PublishSubject
 
@@ -36,5 +35,4 @@ class PermissionEvent(
 
 class BackPressEvent : ViewEvent()
 
-class SuDialogEvent(val policy: Policy) : ViewEvent()
 class DieEvent : ViewEvent()
