--- app/src/main/java/com/topjohnwu/magisk/ui/home/MagiskItem.kt
@@ -1,6 +1,0 @@
-package com.topjohnwu.magisk.ui.home
-
-
-enum class MagiskItem {
-    MANAGER, MAGISK
-}
