--- app/src/main/java/com/topjohnwu/magisk/ui/home/MagiskState.kt
@@ -1,6 +1,0 @@
-package com.topjohnwu.magisk.ui.home
-
-
-enum class MagiskState {
-    NO_ROOT, NOT_INSTALLED, UP_TO_DATE, OBSOLETE, LOADING
-}
