--- app/src/main/java/com/topjohnwu/magisk/ui/home/SafetyNetState.kt
@@ -1,5 +1,0 @@
-package com.topjohnwu.magisk.ui.home
-
-enum class SafetyNetState {
-    LOADING, PASS, FAILED, IDLE
-}
