--- app/src/main/java/com/topjohnwu/magisk/utils/feature/Annotations.kt
@@ -1,16 +1,0 @@
-package com.topjohnwu.magisk.utils.feature
-
-@Retention(AnnotationRetention.SOURCE)
-annotation class WIP
-
-
-@Retention(AnnotationRetention.SOURCE)
-annotation class Beta
-
-
-@Retention(AnnotationRetention.SOURCE)
-annotation class Alpha
-
-
-@Retention(AnnotationRetention.SOURCE)
-annotation class Unstable
